rename soome test class name
