+++ hugegraph-test/src/main/java/com/baidu/hugegraph/unit/UnitTestSuite.java
@@ -22,7 +22,6 @@ package com.baidu.hugegraph.unit;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
-import com.baidu.hugegraph.unit.util.RateLimiterTest;
 import com.baidu.hugegraph.unit.cache.CacheManagerTest;
 import com.baidu.hugegraph.unit.cache.CacheTest;
 import com.baidu.hugegraph.unit.cache.CachedGraphTransactionTest;
@@ -37,11 +36,8 @@ import com.baidu.hugegraph.unit.core.ConditionTest;
 import com.baidu.hugegraph.unit.core.DataTypeTest;
 import com.baidu.hugegraph.unit.core.DirectionsTest;
 import com.baidu.hugegraph.unit.core.ExceptionTest;
-import com.baidu.hugegraph.unit.core.IdSetTest;
-import com.baidu.hugegraph.unit.core.Int2IntsMapTest;
 import com.baidu.hugegraph.unit.core.LocksTableTest;
 import com.baidu.hugegraph.unit.core.PageStateTest;
-import com.baidu.hugegraph.unit.core.ObjectIntMappingTest;
 import com.baidu.hugegraph.unit.core.QueryTest;
 import com.baidu.hugegraph.unit.core.RangeTest;
 import com.baidu.hugegraph.unit.core.RolePermissionTest;
@@ -66,11 +62,17 @@ import com.baidu.hugegraph.unit.serializer.SerializerFactoryTest;
 import com.baidu.hugegraph.unit.serializer.StoreSerializerTest;
 import com.baidu.hugegraph.unit.serializer.TableBackendEntryTest;
 import com.baidu.hugegraph.unit.serializer.TextBackendEntryTest;
-import com.baidu.hugegraph.unit.util.CollectionFactoryTest;
 import com.baidu.hugegraph.unit.util.CompressUtilTest;
 import com.baidu.hugegraph.unit.util.JsonUtilTest;
+import com.baidu.hugegraph.unit.util.RateLimiterTest;
 import com.baidu.hugegraph.unit.util.StringEncodingTest;
 import com.baidu.hugegraph.unit.util.VersionTest;
+import com.baidu.hugegraph.unit.util.collection.CollectionFactoryTest;
+import com.baidu.hugegraph.unit.util.collection.IdSetTest;
+import com.baidu.hugegraph.unit.util.collection.Int2IntsMapTest;
+import com.baidu.hugegraph.unit.util.collection.IntMapTest;
+import com.baidu.hugegraph.unit.util.collection.IntSetTest;
+import com.baidu.hugegraph.unit.util.collection.ObjectIntMappingTest;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
@@ -109,9 +111,6 @@ import com.baidu.hugegraph.unit.util.VersionTest;
     BackendStoreSystemInfoTest.class,
     TraversalUtilTest.class,
     PageStateTest.class,
-    Int2IntsMapTest.class,
-    ObjectIntMappingTest.class,
-    IdSetTest.class,
 
     /* serializer */
     BytesBufferTest.class,
@@ -140,9 +139,16 @@ import com.baidu.hugegraph.unit.util.VersionTest;
     JsonUtilTest.class,
     StringEncodingTest.class,
     CompressUtilTest.class,
-    CollectionFactoryTest.class,
     RateLimiterTest.FixedTimerWindowRateLimiterTest.class,
-    RateLimiterTest.FixedWatchWindowRateLimiterTest.class
+    RateLimiterTest.FixedWatchWindowRateLimiterTest.class,
+
+    /* utils.collection */
+    CollectionFactoryTest.class,
+    Int2IntsMapTest.class,
+    ObjectIntMappingTest.class,
+    IdSetTest.class,
+    IntMapTest.class,
+    IntSetTest.class
 })
 public class UnitTestSuite {
 }
