+++ hugegraph-test/src/main/java/com/baidu/hugegraph/unit/util/collection/Int2IntsMapTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.baidu.hugegraph.unit.core;
+package com.baidu.hugegraph.unit.util.collection;
 
 import java.util.HashSet;
 import java.util.Random;
