+++ hugegraph-test/src/main/java/com/baidu/hugegraph/unit/util/collection/IntSetTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.baidu.hugegraph.unit.util;
+package com.baidu.hugegraph.unit.util.collection;
 
 import java.util.Random;
