+++ hugegraph-test/src/main/java/com/baidu/hugegraph/unit/util/collection/ObjectIntMappingTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.baidu.hugegraph.unit.core;
+package com.baidu.hugegraph.unit.util.collection;
 
 import java.util.Iterator;
 import java.util.LinkedHashSet;
